+++ tests/ui/test_attr_strategy_on_pattern.rs
@@ -4,8 +4,8 @@ fn my_gen(_g: &mut quickcheck::Gen) -> i32 {
   0
 }
 
-#[quickcheck(z = "my_gen")]
-fn t(a: i32, b: String) -> bool {
+#[quickcheck]
+fn t(#[strategy(my_gen)] (a, b): (i32, i32)) -> bool {
   let _ = (a, b);
   true
 }